requirements.txt
streamlit
transformers
torch
faiss-cpu
pandas
numpy
pillow
openai
